Plane: split main servo output functions into a separate file
